# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "User Authentication API"
  description: "APIs for user registration and login with password policy checks"
basePath: "/"
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /getFilterTypes:
    get:
      tags:
        - ALL
      summary:  activity
      description: Logs .
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: PerformedBy
          in: query
          required: false
          type: string
          description: ID of the person performing the action (defaults to '2' if not provided)
      responses:
        200:
          description: Activity inserted successfully
          schema:
            type: object
            properties:
              message:
                type: string
                example: Activity inserted successfully
              status:
                type: boolean
                example: true
        400:
          description: Missing required fields
        500:
          description: Failed to insert activity
...
  /api/AddAHUFilter:
    post:
      summary: "Add AHU Filter"
      description: "Adds a new AHU filter. FilterId, Barcode, and Lable are auto-generated."
      tags:
        - Filter
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - AHUId
              - Lable
              - AssetType
              - Manufacturer
              - InstalledOn
              - Location
              - Dimensions
              - MicronRating
              - CleaningLimit
              - LastCleaningDate
              - ValidOperationLife
              - CurrentStatus
              - AvailabilityStatus
              - Specifications
              - CreatedBy
            properties:
              AHUId:
                type: integer
                example: 2
              Lable:
                type: string
                example: "A100"
              AssetType:
                type: integer
                example: 2
              Manufacturer:
                type: string
                example: "Blue Star"
              InstalledOn:
                type: string
                format: date
                example: "2024-06-15"
              Location:
                type: integer
                example: 2
              Dimensions:
                type: string
                example: "24x12x2"
              MicronRating:
                type: string
                example: "10 Micron"
              CleaningLimit:
                type: string
                example: "5 cleanings"
              LastCleaningDate:
                type: string
                format: date
                example: "2024-07-01"
              ValidOperationLife:
                type: integer
                example: 365
              CurrentStatus:
                type: string
                example: "Active"
              AvailabilityStatus:
                type: string
                example: "Available"
              Specifications:
                type: string
                example: "HEPA filter"
              CreatedBy:
                type: integer
                example: 1001
      responses:
        200:
          description: "Asset item inserted successfully"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: "Asset item inserted successfully"
        500:
          description: "Server error"
  /api/getAHUIds:
    get:
      summary: "Get AHU IDs"
      description: "Returns a list of AHU IDs from the system."
      tags:
        - Filter
      responses:
        200:
          description: "List of AHU IDs"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    AHUId:
                      type: integer
                      example: 5
                    AHUName:
                      type: string
                      example: "AHU Block A"
        500:
          description: "Server error"
  /api/getFilterType:
    get:
      summary: "Get Filter Types"
      description: "Returns a list of available filter types"
      tags:
        - Filter
      responses:
        200:
          description: "List of filter types"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    FilterTypeId:
                      type: integer
                      example: 1
                    FilterTypeName:
                      type: string
                      example: "HEPA Filter"
        500:
          description: "Server error"
  /api/getLocationType:
    get:
      summary: "Get Location Types"
      description: "Returns a list of location types"
      tags:
        - Filter
      responses:
        200:
          description: "List of location types"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    LocationTypeId:
                      type: integer
                      example: 1
                    LocationTypeName:
                      type: string
                      example: "Plant Room"
        500:
          description: "Server error"
  /api/getFiltersList:
    get:
      summary: Get Filters List
      description: Returns a complete list of filters from the system based on optional filters.
      tags:
        - Filter
      produces:
        - application/json
      parameters:
        - name: BlockId
          in: query
          required: true
          type: integer
          description: ID of the block (use 0 for all blocks)
          example: 0
        - name: Status
          in: query
          required: true
          type: integer
          description: Status of the filter (use 0 for all statuses)
          example: 5
        - name: Barcode
          in: query
          required: true
          type: string
          description: Barcode of the filter (use 0 for all)
          example: "0"
      responses:
        200:
          description: Successful response with list of filters
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    FilterId:
                      type: string
                      example: "FLT123456"
                    AHUId:
                      type: integer
                      example: 5
                    Code:
                      type: string
                      example: "A100"
                    Manufacturer:
                      type: string
                      example: "Blue Star"
                    InstalledOn:
                      type: string
                      format: date
                      example: "2024-06-15"
                    Location:
                      type: string
                      example: "Bangalore Plant"
                    Barcode:
                      type: string
                      example: "12345678"
                    Status:
                      type: integer
                      example: 1
        500:
          description: Internal Server Error
  /api/CheckOrderedStage:
    post:
      summary: "Check and Update to Next Ordered Stage"
      description: >
        Checks the current stage of a filter by Barcode. If it's an ordered stage, 
        moves it to the next ordered stage and returns a success or failure message.
      tags:
        - Filter
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - BarcodeNum
              - OrderNum
              - PerformedBy
            properties:
              BarcodeNum:
                type: string
                description: Barcode of the filter
                example: "skjdh767"
              OrderNum:
                type: integer
                description: User's current known stage order
                example: 3
              PerformedBy:
                type: integer
                description: User ID who performed the action
                example: 3
      responses:
        200:
          description: Status of the transition to next ordered stage
          schema:
            type: object
            properties:
              UserGivenOrder:
                type: integer
                example: 3
              ResultMessage:
                type: string
                example: "Status updated to next ordered stage: 4"
              Status:
                type: string
                example: "true"
        500:
          description: Internal Server Error
  /api/UpdateFilterStatus:
    post:
      summary: "Add Filter History Entries (Bulk)"
      description: "Accepts an array of filter history records and inserts them one by one into the FilterHistory table."
      tags:
        - Filter
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              type: object
              required:
                - FilterId
                - StageId
                - Status
                - performedBy
                - Barcode
              properties:
                FilterId:
                  type: integer
                  example: 102
                StageId:
                  type: integer
                  example: 3
                Remarks:
                  type: string
                  example: "Filter cleaned and dried"
                Status:
                  type: integer
                  example: 1
                washingEquipId:
                  type: integer
                  example: 5
                AirPressure:
                  type: string
                  example: "80 PSI"
                RoWaterPressure:
                  type: string
                  example: "60 PSI"
                Duration:
                  type: integer
                  example: 30
                DryerEquipId:
                  type: string
                  example: "DRY-01"
                DryerTemp:
                  type: string
                  example: "70°C"
                performedBy:
                  type: integer
                  example: 1001
                Barcode:
                  type: string
                  example: "BARCODE12345"
                WashingReasonId:
                  type: integer
                  example: 1001
      responses:
        200:
          description: "All filter history entries processed"
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Processed all filter history entries"
              success:
                type: boolean
                example: true
              results:
                type: array
                items:
                  type: object
                  properties:
                    Barcode:
                      type: string
                      example: "BARCODE12345"
                    result:
                      type: object
                      properties:
                        Status:
                          type: boolean
                          example: true
                        message:
                          type: string
                          example: "Inserted successfully"
        500:
          description: "Internal Server Error"
  /api/GetFilterStatus:
    get:
      summary: "Get Filter History"
      description: "Fetches the complete maintenance history of a filter using its FilterId."
      tags:
        - Filter
      produces:
        - application/json
      parameters:
        - name: FilterId
          in: query
          type: integer
          description: ID of the filter to retrieve history for
          example: 101
      responses:
        200:
          description: "Successful response with filter history"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    StageId:
                      type: integer
                      example: 3
                    Remarks:
                      type: string
                      example: "Filter cleaned and dried"
                    Status:
                      type: integer
                      example: 1
                    washingEquipId:
                      type: integer
                      example: 5
                    AirPressure:
                      type: string
                      example: "80 PSI"
                    RoWaterPressure:
                      type: string
                      example: "60 PSI"
                    Duration:
                      type: integer
                      example: 30
                    DryerEquipId:
                      type: string
                      example: "DRY-01"
                    DryerTemp:
                      type: string
                      example: "70°C"
                    CreatedOn:
                      type: string
                      format: date-time
                      example: "2025-07-04T15:30:00"
                    CreatedBy:
                      type: integer
                      example: 1001
                    Barcode:
                      type: string
                      example: "BARCODE12345"
        400:
          description: "FilterId is missing or invalid"
        500:
          description: "Internal Server Error"
  /api/GetMastersEquipment:
    get:
      summary: "Get Master Equipment List"
      description: "Returns the list of equipment based on whether it is a dryer or not."
      tags:
        - Master
      produces:
        - application/json
      parameters:
        - name: Isdryer
          in: query
          required: true
          type: integer
          description: "0 for non-dryer, 1 for dryer"
          example: 0
      responses:
        200:
          description: "List of equipment"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    EquipmentId:
                      type: integer
                      example: 1
                    EquipmentName:
                      type: string
                      example: "Washer 3000"
                    IsDryer:
                      type: boolean
                      example: false
        400:
          description: "Missing or invalid query parameter"
        500:
          description: "Internal Server Error"
  /api/GetMastersReasons:
    get:
      summary: "Get Master Reasons"
      description: "Returns a list of reason masters based on the given type."
      tags:
        - Master
      produces:
        - application/json
      parameters:
        - name: Type
          in: query
          required: true
          type: integer
          description: "Type of reason to retrieve (e.g., 1 for washing reasons, 2 for drying reasons)"
          example: 1
      responses:
        200:
          description: "List of reasons fetched successfully"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    ReasonId:
                      type: integer
                      example: 101
                    ReasonText:
                      type: string
                      example: "Filter damaged during wash"
                    Type:
                      type: integer
                      example: 1
        400:
          description: "Missing or invalid query parameter"
        500:
          description: "Internal Server Error"
  /api/GetMastersPressure:
    get:
      summary: "Get Master Pressure Settings"
      description: "Returns a list of master pressure configurations based on the specified type."
      tags:
        - Master
      produces:
        - application/json
      parameters:
        - name: Type
          in: query
          required: true
          type: integer
          description: "Type of pressure (e.g., 1 for Air, 2 for RO Water)"
          example: 1
      responses:
        200:
          description: "List of pressure settings fetched successfully"
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    PressureId:
                      type: integer
                      example: 201
                    PressureValue:
                      type: string
                      example: "80 PSI"
                    Type:
                      type: integer
                      example: 1
        400:
          description: "Missing or invalid query parameter"
        500:
          description: "Internal Server Error"
  /api/getAllUsers:
    get:
      summary: Get All Users
      description: Returns a list of all users from the system.
      tags:
        - User
      produces:
        - application/json
      responses:
        200:
          description: List of users fetched successfully
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  type: object
                  properties:
                    UserId:
                      type: integer
                      example: 1
                    UserName:
                      type: string
                      example: "john.doe"
                    Email:
                      type: string
                      example: "john.doe@example.com"
                    Role:
                      type: string
                      example: "Admin"
        500:
          description: Internal Server Error
...
  /api/updateAHUstatus:
    post:
      summary: Delete AHU Asset
      description: Deletes an existing Air Handling Unit (AHU) record based on the provided ID.
      tags:
        - AHU
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
              IsActive:
                type: integer
                example: 0

      responses:
        200:
          description: AHU asset deleted successfully
        400:
          description: Bad request (missing or invalid ID)
        404:
          description: AHU asset not found
        500:
          description: Internal server error
...